+++ src/main/java/com/fjnu/intelligentwardrobe/Mapper/UserMapper.java
@@ -8,8 +8,8 @@ import java.util.List;
 
 @Mapper
 public interface UserMapper {
-    List<User> findUser(Integer phone,String password);
-    List<User> findUserByPhone(Integer phone);
-    Integer saveUser(Integer phone,String password);
+    List<User> findUser(Long phone,String password);
+    List<User> findUserByPhone(Long phone);
+    Integer saveUser(Long phone,String password);
     Integer updateUser(Integer uid,String column,String value);
 }
