+++ src/main/java/com/fjnu/intelligentwardrobe/Service/UserService.java
@@ -5,8 +5,8 @@ import com.fjnu.intelligentwardrobe.Bean.User;
 import java.util.List;
 
 public interface UserService {
-    List<User> findUser(Integer phone,String password);
-    List<User> findUserByPhone(Integer phone);
-    boolean saveUser(Integer phone,String password);
+    List<User> findUser(Long phone,String password);
+    List<User> findUserByPhone(Long phone);
+    boolean saveUser(Long phone,String password);
     boolean updateUser(Integer uid,String column,String value);
 }
